refactor mount objects
Use only runtime::Mount instead of the polymorphic hierarchy
(runtime::UserMount + runtime::SiteMount).

The difference between UserMount and SiteMount was that UserMount
was performing an additional security check to ensure that the
user was not trying to mount a directory not visible/accessible
to him on the host. Now we just perform such check for both types
of mounts.