fix handling of non-ascii characters
Since we switched to musl, the non-ascii characters were not
handled correctly anymore. E.g. the docker image
ethcscs/dockerfiles:image-with-non-ascii-characters would not
contain the expected file and folder names.

We needed to explicitely set the locale with std::setlocale()
in order to enable musl to properly handle non-ascii characters.