Run SSH hook tests with both musl-only and glibc-only images
Run tests in musl-only containers as well as glib-only containers, in order to make
sure that the custom OpenSSH injected into the container is standalone (as expected)
and doesn't have any dependency on musl or glibc. It happened in the past that the
custom sshd depended on the NSS dynamic libraries of glibc, so sshd used to
break in containers without glibc.